Stop using regular expressions for the badwords test.
Converting the badwords strings in the config file to regular
expressions means that we should really be escaping the strings for
conversion to a regex, and that seems like a lot of trouble. The regular
expressions were only used to add word-boundary checks, which, besides
an undocumented GNU extension, is pretty easy to do without regexes.