Rename the get_suite function
Nothing actually happens with the test suite object returned by
get_suite; the call to CU_add_suite both creates the test suite and adds
it to the test registry. The return value is only used to determine
whether the suite was successfully created.

Rename get_suite to add_test_suites and have it return a boolean for
success/failure. This makes it more clear that registering the suite is
a side-effect of the function, and allows for multiple test suites per
.c file.