WIP Add functions to mock rpm header operations
This will allow for the creation of unit tests for the functions that
act upon rpm header data without needing to create fake rpms for each
test.

Add wrapper functions for headerNew, headerFree, headerLink, and
headerGet, and add a support function to set the return value for
RPMTAG_FILENAMES. More tag types can be added as needed.